Comment the nameless addresses
Addresses proper are considered to be "nameless".

You can have a POI with a name and an address. In that case name will have the POI name and housenumber and street will still be set.

That largely comes from the OSM data model where we happily mix POI and address data. And where an address may only appear together with a POI.